Compose: Rework types of createHigherOrderComponent for closer match to API (#37795)
The existing types in `createHigherOrderComponent<TOuter, TInner>` are somewhat
_surprising_ because they require defining what possible prop types are allowed
for components wrapped by the higher order component it produces. Additionally
it doesn't concern itself with preventing the wrapped component from accepting
or demanding props it itself provides.

In this patch we're reworking the types to only require a single type parameter,
which is the type of props that the higher order component injects into the
wrapped component. The prop types for the wrapped component are inferred only
when wrapping a component, and the injected props are removed from the external
interface to avoid confusion.