Disallow setting in-progress or failed uploads as featured
In-progress or failed image uploads cannot be set as the featured image
as the media does not exist on the server. Relying upon a negative ID
number did not properly disable in-progress uploads, as negative IDs are
a iOS-only coincidence due to hashing of the integer ID.

These changes rely upon media upload sync events and the presence of an
`image` prop to determine when the upload is completed successfully.