Add basic tests using Playwright
This includes the tests for visiting admin page and login into WordPress.

The new tests are created in a new file that are suffixed with `-pw`.

For the user login test, we can get rid of `pressKeyWithModifier` that was used to simulate "Select All" behaviour. This was needed with Puppeteer, but Playwright provides this natively.