Refactor Site Editor test utils to the `e2e-test-utils` package (#38463)
* Refactor site editor utils from `e2e-tests` to `e2e-test-utils`

* Refactor `toggleMoreMenu` to allow for both site and post editors

* Abstract `GutenbergContext` into own type

* Refactor `clickOnMoreMenuItem` to accept editor context

* Refactor `getEditedPostContent` into own function

That didn't seem to be used anywhere. Perhaps a bit of a red flag.
Also decided not to go with adding an optional parameter to the
existing `getEditedPostContent` because the two functions differed
significantly. Perhaps it is not very consistent, however.

* Rename `getEditedPostContent` for site editor and replace instances

Function renamed to `getEditedPageContent` to avoid conflicts with
existing functions. Also found instances in which it was used and
replaced them.

* Refactor `disableWelcomeGuide` into own function

* Refactor Site Editor navigation panel utility functions into own functions

* Add an extra option to `toggleMoreMenu` and fix tests

The function can now wait for opening or closing as required, making the clicking
of menu items more stable.

* Fix more tools menu content selector

* Fix site editor more tools menu content selector

* Rename some functions, disable welcome guide by default

* Rename getEditedPageContent to getCurrentSiteEditorContent

* Rename disableWelcomeGuide function