Links changed for Develeoper.wordpress.org (#38841)
* Links changed for Develeoper.wordpress.org

On this page: https://developer.wordpress.org/block-editor/how-to-guides/data-basics/
The links in the page body go 404. The .md needed to be removed. 
This results in the links not working when looking on it on the GitHub repo though. 
I feel though, having them not work on the repo is less misleading than the links not working on the final destination of the tutorial 
I am aware that I changed it a few days ago into the other direction. Back then I wasn't aware about this conversion problems from GitHub to developer.wordpress.org

* Update README.md

Co-authored-by: Greg Ziółkowski <grzegorz@gziolo.pl>