[data]: Fill out minimal type definition for DataRegistry
The `DataRegistry` type has been a hollow shell of its real shape.
In this patch we're adding basic types for all the methods and
properties it exposes to the app.

There is value to glean from expanding these types but as a starter
this update should prevent someone from getting a false type error
when attempting basic use of the registry in typed code.