Build: Use temporarily non-tagged node module project dependencies
After updating all node module forked libraries to
'Gradle 7.4 & AGP 7.1.1', this commit update the node module project
dependencies to point to that 'update/gradle-to-7.3.3-agp-to-7.0.4'
branch for all of them, which has this new version.

This is mainly done for testing purposes. Afterwards, and upon
verifying that everything is okay, all those draft node module forked
library PRs will get merged and a new node module forked library version
will be tagged and used in favor of these temporarily coordinate