Docs: Emphasized requirement to enclose in quotation marks wp-env run commands with options
The help text for the wp-env run command states this requirement. For emphasis, I added an introduction to the run command that re-states the requirement.

Also, for this existing example to be consistent with the previous run command examples, 'Installing a plugin or theme on the development instance', I replaced 'wp-env run cli plugin install' with 'wp-env run cli wp plugin install'.