Core Data: Rename `types` directory to `entity-types`
When we introduced the types for the core entities in #38666 we used a
directory named `types`. Unfortunately this meant that we can't create the
customary file `src/types.ts` because of the name clash in module resolution.

In this patch we're renaming the directory so that we can separately
import `entity-types(/index.ts)` and also `types(.ts)` without the
naming conflict.