Core Data: Pass explicit `undefined` initial value to `createContext`
Part of #39211

Previously we have been calling `createContext()` without any
arguments, but there's a funny note in the official React type
for that function.

> // If you thought this should be optional, see
> // https://github.com/DefinitelyTyped/DefinitelyTyped/pull/24509#issuecomment-382213106

Although not passing an argument is practically the same as passing
`undefined` as the argument we have a type error that TypeScript
doesn't like while we're relying on it to parse JS files with the
JSDoc typings.

In this patch we're just adding the explicit `undefined` which should
have no behavioral change on the output but removes the type issue.