Core data: create type for undo state and initialize atomically
Part of #39211

In this patch we're adding some types to the undo state tracked
in the core data store. Additionally we're intializing it in one
atomic operation whereas previously that took two partial assignments.

Futher, a JSDoc comment has been moved from its previously erroneous
location and back to the function it describes. This error had been
overlooked in a previous change to the module.

The purpose is to remove type issues when toggling on TypeScript
across the `core-data` package.

This is almost entirely a type-level change but the initialization
of `UNDO_INITIAL_STATE` has been collapsed into a single assignment.

Verify that the editor undo/redo functionality works as expected.
Given that the one code change impacts the initialization it should
be quickly obvious if there's a problem or regression with undo/redo.