WIP: Block Parser: Unserialized parsed JSON attributes to avoid corruption.
Previously we've been parsing the raw JSON attributes in a block comment
delimiter and sending them into the system. This mostly works because the
transformations we make before serializing are semantically equivalent
in display contexts.

However, if a block wants to store attributes that have more exacting
equality requirements this may lead to loading different states than were
saved, for example, when saving a CSS class name in an attribute.

In this patch we're closing the loop on save and load so that if you save
content with attribute values that get transformed then you will get
identical values back out upon load.