Add useSelectors hook for getting a store's selectors
The `useSelect` function is used to cover 2 use cases, switching
between them with the first parameter's type.

The original use case is to fetch data during render. The 2nd use case
simply returns a store's selectors, so they can be called in event
handlers.

This 2nd use case does not need most of the hooks the original use is
calling. However because of how hooks work, they needed to be called
anyway in this path.

In the new function, the only remaining hook call for getting the
store's selectors is to `useRegistry`.

It also allows to simplify useSelect, which doesn't need any of the
checks on the type of argument anymore, and can simply return the map
output.

I didn't test the impact of not having to call
these hooks, but I guess it's not 0. Especially if many components use
this.

For now `useSelect` preserves its double function, because it is
exported as a public API. In a next step it's possibel to internally
already use a simpler form that does 1 thing, and keep a copy of the old
one only for exporting the API before it's removed.