Core data: Expand type signature of useQuerySelect (#39656)
Part of #39211

In this patch we're expanding the type signature of `useQuerySelect`
and fixing a minor type-related issue that was formerly there.

We want to do great things with the `core-data` type system but all of these
little nuisances get in the way when we're trying to study those things in
isolation. This is a preparatory change to eliminate some of the noise in the
existing types.

As a type-only change this will have no impact on the built code.
Please audit the types and the use of the wrapper `EnrichedSelectors`.