Prevent expanded button displaying in some cases
As reported here, the new expanded button was displaying in some cases when the inserter button was tapped: https://github.com/WordPress/gutenberg/pull/39726#issuecomment-1091752804

This was happening when the default block was selected. This is because the default block is the only block type that's replaced (and therefore de-selected) as part of the inserter flow.

With this commit, we now check for whether the default block is being replaced (if so, it's value will be "null"). If it is, the value of "isDefaultView" is set to "false".