Switch to preset approach for named colors
This change is approach was driven by a desire to not add any new block support related attributes and css classes. The prior implementation added a lot of additional CSS classes, bloated the block class lists, and would have needed migration and backwards compatibility support.

This new approach relies on being able to add inline styles with `var()` calls to use preset colors so switching themes still updates colors as expected and without the need to update blocks.

This approach is intended to be temporary until the new styles engine implements a cleaner result and the preset vars are required to be stored in the border color properties.