Turn 'isDefaultView' into 'global' editor setting
It has become necessary to change the value of 'isDefaultView' at various levels of components, which is difficult to do with the existing implementation (i.e. it's currently defined in a parent component, making it difficult to update its value from within a child component).

With this commit, 'isDefaultView' is converted into a block editor setting, similar to other settings like 'hasFixedToolbar', 'focusMode', 'hasReducedUI', etc.

Following this refactor, it will be easier to change the value of 'isDefaultView' at various component levels. It will also make it easier to iterate on the usage of the value in the future.