Simplify useSelect by dispatching to useSelectors
This is a deliberate rules-of-hooks violation that is safe to do
because the condition stays same for each invocation, mapSelect's type
never changes.

Unfortunately it meant disabling the rules-of-hooks check in the whole
function body, but there's no getting around that. However the chance
for the rule to be needed in this function is rather small.