Media & Text: Remove font size declaration from template
This represents and accessibility improvement to this block. WCAG 2.1 (Level A) Success Criterion 1.3.1 Info and Relationships states that “[…] nformation and relationships that are implied by visual or auditory formatting are preserved when the presentation format changes”. By making the paragraph a larger font size, the block is effectively meeting the requirements of Failure F2 by emphasizing text without using semantic markup to convey the visual significance of the text to users of assistive technology.

Additionally, if a developer has removed custom font sizes from the Paragraph block, or the site entirely via theme.json, the custom font size is still applied, with no way to remove it.