Remove unnecessary code in web-only file
Following the separation of native-specific code to its own file, this commit further refines some unnecessary code that was left behind in the (now) web-only file.

Specifically, the logic surrounding the 'isCollapsed' prop previously included a check for whether 'isToolbar' is true. Now that there's a native-specific file, 'isCollapsed' will only ever be passed when 'isToolbar' is true. So, the extra check is redundant.