Fix `recommended` ESLint preset when `prettier` is not installed
The `prettier` package is incorrectly mandatory when using the `@wordpress/eslint-plugin/recommended` preset. Failure to install it results in an error.

`prettier` is intended to be an optional peer dependency, with related rulesets added only when the user has installed `prettier`. However, the `@worpress/prettier-config` package eagerly loads `prettier`'s `package.json`, leading to an error when it is not installed.

This change moves the require of `@worpress/prettier-config` inside the relevant `if` block, so that it is only loaded when `prettier` is confirmed to be installed.