Add no comments count placeholders (#40728)
We have both singular and plural editing on the comment title. Also, we have an option for showing comments count or not.
In singular, this last option was not taken into account, showing always the same placeholder. 'One response'

Also, to keep consistency, we change uppercase or lowercase on R letter of Response depending on we are enabling or disabling the comments count.