DOM: isNumberInput: Don't check for non-zero value
Perhaps due to an excess of zeal, `isNumberInput` expected elements to
have a truthy value, which meant that a number input with value=0 would
be rejected by the function.

The correct condition might have been to check that `valueAsNumber` is
not a `NaN` but a proper number. However, I still find this too
restrictive. A number input that hasn't yet received a value from the
user should still be considered a number input. Hence, the condition
that type=number should be enough.