Replace unbounded query to fix gallery settings
Requests using "per_page=-1" (unbounded queries) are not supported on mobile, as per: https://github.com/wordpress-mobile/gutenberg-mobile/issues/2661

As an unbounded query was being used to fetch inner images in the gallery, the "link to" settings were not working as expected on native.

With this commit, the unbounded query is replaced with "imageIds.length" to fix the issue on native.