DOM: Deprecate `isNumberInput` and fix its handling of value=0 (#40896)
* DOM: isNumberInput: Don't check for non-zero value

Perhaps due to an excess of zeal, `isNumberInput` expected elements to
have a truthy value, which meant that a number input with value=0 would
be rejected by the function.

The correct condition might have been to check that `valueAsNumber` is
not a `NaN` but a proper number. However, I still find this too
restrictive. A number input that hasn't yet received a value from the
user should still be considered a number input. Hence, the condition
that type=number should be enough.

* Deprecate `isNumberInput`

Reintroduce the check for valueAsNumber, but check for NaN rather than
falsey values. Since we are deprecating the function, it seems
preferable to not change its behaviour too much, in case third-party
consumers of `@wordpress/dom` are relying on these quirks.

* DOM: Changelog: Add deprecation of isNumberInput

* Update packages/dom/CHANGELOG.md

Co-authored-by: Greg Ziółkowski <grzegorz@gziolo.pl>

* Update packages/dom/CHANGELOG.md

Co-authored-by: Greg Ziółkowski <grzegorz@gziolo.pl>

Co-authored-by: Greg Ziółkowski <grzegorz@gziolo.pl>