[compose]: Fix type for `createHigherOrderComponent`
When refactoring `createHigherOrderComponent` in #37795
an incorrect type was introduced in an attempt to correct
the previous type. The fix in that patch was focused on
the narrow use-case of a _prop-injecting higher-order component_.

In this patch we're lifting the type signature into the function
which calls `createHigherOrderComponent`. There's a limitiation
I'm not sure how to get past when adding the type parameters to
the generalized interface; that is, the function _produced_ by
`createHigherOrderComponent` needs its own type parameters for
input and output specified by the given `mapComponent` mapper.

This change require more type juggling when defining a new
higher-order component but removes the arbitrary constraint
on what props may be passed into or come out of the higher-order
component.