Reorganize global styles menu
Implements the second alternative shown in https://github.com/WordPress/gutenberg/issues/40478
to reorganize the Global Styles panel to keep both Blocks and Browse
styles links in line with the rest of the menu.

Let's iterate over this PR to decide on which of the two options makes
more sense.