Add supports to basic sample code.
Like `attributes` and `save`, `supports` are not automatically inherited from the current version, so they need to be defined in some cases, but the basic sample code does not describe them, so it is difficult for beginners to understand.
I think this should add a description for clarity.

By the way, this page uses const to refer to the global wp object, but other pages use import statements. I feel that mixing the two could cause confusion, but should this page also be an import statement?