Pasting: Fix performance regression due to removeWindowsFragments
Fixes #41826.

removeWindowsFragments was running two String#replace with regular
expressions made computationally expensive due to the use of `.*` and
the `s / dotAll` flag, resulting in severe performance degradations when
handling larger strings of HTML.

The solution is to manually trim the strings via a combination of
String#indexOf and String#substring.