test: Reinstate image caption class name for Image block test fixture
The native editor test fixtures includes a Gallery block v1 (which needs
to upgraded to v2 at some point). Adding a `wp-element-caption` to this
test fixture causes parsing failure.

The lack of `wp-element-caption` class name for the Image block causes
an unnecessary block update during test runs.

The `wp-element-caption` class name was added to the native editor test
fixtures for both the Gallery and Image. The addition was only needed
for the latter. The previous commit erroneously removed the class name
for both. This commit reinstates the class name for the latter.