Create a new handbook page for curating the editor experience (#42022)
* Create a new handbook page for curating the editor experience

This proposes a v1 of a new page that compiles various methods for curating the editor experience.

* Update manifest.json to include new page

* Update toc.json for new handbook page

* Add screenshots to assets folder

* Add example video to assets

Had to speed this video up in order for it to be small enough to include.

* Updating asset references

I can't seem to figure out why the video won't properly render. When I look at the individual file, it seems the size is still too big to display. If I don't add it to the assets file and just drop it in, it works though.

* Rebuilds docs manifest.

* Updating assets to refer to path to trunk rather than my PR

* Resolve conflict with manifest.json

Not sure if this is the right way to resolve a conflict with trunk but this is my attempt :)

* Minor spacing update to manifest.json

* Rebuild the docs.

Co-authored-by: Ryan Welcher <ryan.welcher@automattic.com>