fix: Revert utilization of onSelect from useBlockTypesState
The block created from within this callback appears to be in a format
that is incompatible with the native editor at this time. Specifically,
this appeared to break Embed block variants, e.g. Vimeo. A future
improvement would be to investigate this further in hopes of aligning
web and native.