fix: Correct alt text footnote spacing
Add a space between the two sentences of the alt text footnote.

A space was lacking between the two sentences within the alt text
footnote. This occurred due to the way React parses the separate link
element found within this particular footnote. Originally an appropriate
space was included in the first sentence, but it was accidentally
removed entirely within #38115 to avoid i18n script errors.