Pasting: Dismiss pasted image if file:// schema detected
Fixes #42783

This isn't the fix that I would like, and I am not sure that it should
be merged.

IMO, the underlying issue is that Google Chrome (and Chromium
derivatives) decides to attach to the clipboard data the fallback
snapshot provided by Microsoft Word. Other browsers ignore that data,
only passing attachments of type `text/plain`, `text/html`, and
sometimes `text/rtf`.