fix: Latest Posts categories settings (#43685)
Add a named export to `SelectControl` so the native module aligns with
the web module. The use of a named export was added to code shared
between web and native. Because the latter only defined a default
export, an error was thrown.

- Named export added: https://github.com/WordPress/gutenberg/pull/40737
- Named export used: https://github.com/WordPress/gutenberg/pull/41536

I am uncertain as to why a named export was added originally. My
presumption is that it was recently used so that the import was a local
reference, instead of a importing the module's top-level entry file.