Set font family done after printing
Fixes an issue where the `WP_Webfonts::$to_do` has each font
 family, but the `WP_Webfonts::$done` does not.

Found when working on the `query()`.

After printing the variations, the font family for these printed
variations will be added to the `WP_Webfonts::$done` property.

In order to make this work, a `is_font_family` flag is added
to `extra` data.