Create a new handbook page for propagating updates (#43538)
* Create a new handbook page for propagating updates

Similar to https://github.com/WordPress/gutenberg/pull/42022, this proposes a v1 of a new page that discusses how best to manage propagating updates to your site. The aim is for an initial version to be shared that can then evolve over time to be more comprehensive as needed. Assistance from @justintadlock @gziolo @juanmaguitar @ryanwelcher. 

* Update manifest.json for new doc

* Update toc.json for new handbook page

* Text updates

* Updating the blocks section

Updating based on great feedback from @gziolo.

* Updating bolding and headers

Updating the structure of the doc in light of @juanmaguitar's feedback.

* Updating title of page

* Updating manifest.json with new title

* Rebuild docs.

Co-authored-by: Ryan Welcher <me@ryanwelcher.com>