Add necessary space between text and URL
As we'll always need a space between text and the URL that's returned by the HelpText component, this commit goes ahead to add the space within the component itself. This will make the component easier to use "out of the box", as it won't be necessary for developers to remember.