Split string into multi-line, move tests into isolated runner
The tests verifying behaviors dependent on whether `WP_DEBUG` is
set need to run in an isolated environment so that the constant's
definition doesn't bleed into other tests. The isolated runner is
very slow in comparison to the non-isolated runner so only those
tests dependent on the constant are moved over.