QA: no control structures with empty body
A control structure with an empty body is often indicative of "missing code", i.e. code which should have been added, but was forgotten.

For this reason, empty bodies for a control structure are flagged when using `WordPress-Extra`.

For the control structures in this PR, the "body" part is not needed and PHP allows for that without issue (for select control structures). See: https://3v4l.org/ADFk5

To maintain the descriptiveness of the code, however, `continue` has been added instead.