PHPCS: use PHPCompatibilityWP
The current PHPCS ruleset `phpcs.xml.dist` uses the `PHPCompatibility` ruleset to flag code which could be incompatible with the supported PHP versions.

For code used in the Context of WordPress, a dedicated [`PHPCompatibilityWP`](https://github.com/PHPCompatibility/PHPCompatibilityWP) ruleset is available to prevent PHPCompatibility flagging code for which WordPress prevents polyfills.

Fixes 44541