Tag Processor: throw when supplied unacceptible attribute names. (#44431)
The `WP_HTML_Tag_Processor` allows setting new HTML attributes with
a given name and value. Previously this has allowed any string input
for the attribute name, but we have to be careful not to print output
that might break the HTML we're modifying.

In this patch we're adding a check against the given attribute name
and rejecting invalid or unacceptible names. WordPress here is more
restrictive than HTML5.

In order to avoid crashing real sites this only throws an exception when
`WP_DEBUG` is set and active; in production environments it ignores
the attribute update, skipping the invalid name.

Co-authored-by: Adam Zieliński <adam@adamziel.com>