Perf Tests: Refactor use of ENV and git in test setup
In this patch we're refactoring how the performance tests are
configured, relying on the ENV values set by Github Actions and adding
new semantic terms for existing implicit conditions.

Additionally, the custom `git` abstraction has been removed and
replaced with `SimpleGit` for a more concrete represetnation of
what is being performed during the test setup. This change will
enable more complicated uses of `git` for optimizing the run script,
and by moving out of the abstraction it will minimize any changes
based on that kind of `git` usage.

There should be no functional changes in this patch and only the
code surrounding test setup should change. The term "branch" is
replaced in a few places with "ref" since this script is designed
to accept any `git` ref pointing to commits, and in fact has been
in daily use with a mix of tags, commit SHAs, and branch names.