Perf tests: checkout first from the local repo before downloading
When running the performance tests in Github Actions we start with a
local shallow clone of the Gutenberg repo from a previous step.
Previously we've been cloning a new copy of the repo from Github
directly.

In this patch we're starting each checkout by looking first in the
local clone if it's already available. If not then we'll reach out
to the remote repo on Github.