fix: Hide keyboard button forcibly dismisses the virtual keyboard
Due to race conditions between Aztec focus effects and React Native
gesture handlers, it is possible to select one block while placing focus
in a different block's text input. If the actively select block is not a
rich text field, tapping the dismiss keyboard button will not dismiss
the keyboard via clearing the selected block, as a non-rich-text block
does not trigger a blur event. So, this change explicitly dismisses the
keyboard rather than relying upon the selected block clear.

This is a hack. Ideally, the Aztec call loop is improved so that it is
not possible to select/focus more than one block.