Fix: Prevent infinite loop in Tag Processor in certain truncated documents.
Previously the Tag Processor class has performed unchecked arithmetic on
the result of `strpos()` when looking to close out HTML comments and other
special sections (CDATA, doctype-declaration). This led to a situation
where by means of type-coercion we added an integer value to a `false`
returned by the string function, setting the cursor in the tag processor
to a low index in the document, and creating an infinite loop condition.

In this patch we're checking the results of calling `strpos()` in those
places to avoid the type error and abort from the processor if we fail
to find the end of those associated document sections.