Add styles to <figure/> element in core/image block during rendering of a gallery
When a gallery image contains a border radius and captions, the caption
container covers up the rounded corners of an image.

This commit checks to see if the border radius is set on core/image,
and if so, applies styles to the <figure /> element that will allow
captions to render properly.