Alphabetically sort blocks in global styles
The blocks in global styles firstly list the core blocks in an
undefined order, and then the non-core blocks also in an undefined
order. This change lists the core blocks alphabetically by title, and
then non-core blocks alphabetically by title.

This gives a more predictable order of blocks when using global styles.