Test: Skip performing tests in PRs if merge-conflict exists for target branch.
From time to time while developing a PR a merge conflict appears between the
branch in the PR and the target branch against which it's based. In these cases
all of our tests continue to run, but they run against an old ref in the target
branch and the results of the tests won't reflect any possible merge of the PR.
That is, if the tests were to pass, they would only represent passing if the
newer commits on the target branch hadn't occurred.

In this patch we're adding a step at the beginning of the test run which checks
indirectly if we have a merge conflict by looking at the first parent of the PR's
merge branch tip and checking if it's the same as base ref tip.

If it's a different commit then we can infer that Github hasn't updated the PR's
merge branch since that time due to a conflict, and we can skip running the tests.