Add placeholder to Page List block
When there are no pages, the Page List block was showing an
unnecessary notice when it was a child of a Navigation block.

In order to more closely match the rendered front end and reduce
clutter, I instead added a placeholder so that users whose
intent is to have no pages can see a better representation
of their site.