Perf Tests: Fix test build process to call wp-scripts
In #45284 we started skipping the generate-types phase of the build
when running the performance tests. This was done to avoid the time
it takes to generate those types when they are ignored by the rest
of the test suite.

In that patch we accidentally skipped running `wp-scripts build`,
and in this patch we're bringing that back to fix the mistake.