Perf Tests: Run each test variation in a separate OS process
We've been avoiding running the tests in parallel to avoid interference
between the threads impacting the results.

However we've discovered that such interference already exists and likely
stems from utilization of the test runner container for other workflows,
possibly from other projects.

In this patch we're forking when running the tests to see if we can get
better CPU utilization and finish tests faster without sacrificing data
quality.