Move post/page title to top bar
The title's position in the editor can be
inaccurate depending on how themes are styled;
to more accurately represent how a post will
look when published, this commit moves the
title from the editor body to the top toolbar.

This mirrors the treatment for the toolbar in
Full Site Editing. For reference, see edit-site/
header-edit-mode/document-actions.