Tag Processor: Add bookmark system for tracking semantic locations in document
It can be helpful to track a location in an HTML document while updates
are being made to it such that we can instruct the Tag Processor to seek
to the location of one of the bookmarks.

In this patch we're introducing a bookmarks system to do just that.
Bookmarks are referenced by name and handled internally by a tracking
object which will follow all updates made to the document. It will be
possible to rewind or jump around a document by setting a bookmark and
then calling `seek( $bookmark_name )` to move there.

Co-authored-by: Adam Zielinski <adam@adamziel.com>
Co-authored-by: Dennis Snell <dennis.snell@automattic.com>