display fontSize short name in FontSizePicker if preset defines it
If a fontSize preset defines a shortName property and the property is between 1 and 3 characters long this shortName will get displayed instead of the tshirt size value.

This of course only applies if there are fewer than 6 font size presets were defined in which case the FontSizePicker shows the full name in the SelectControl instead.