Add animationSettings to Navigator; refactor NavigationBlockInspector
I added the option to override the default animation settings
in the Navigator to allow for customization of the user experience.

In the BlockInspector, we now calculate the depth of the
selected block and, based on that information, configure
the animation settings to give the illusion of a user
drilling down into the hierarchy.

Importantly, we do NOT execute an animation when the
NavigationInspector is first loaded or when we select
new blocka at the same depth.

Currently, we only animate when selecting any descendant
of the root Navigation block, or when we select the Navigation
block after previously having selected a descendant.