Table: Display fixed width table cells option by default
This has the effect of the table displaying the settings table consistently regardless of whether it is in the placeholder state or not. The Header and Footer settings were omitted as they really don't make a lot of sense while there is no content for the table and could lead to an empty table with only the header and footer.