HTML API: Move into 6.2 Compat Folder since inclusion in Core
Since the merge of the Tag Processor into Core in [55203][55203]
it's no longer needed in Gutenberg in the experimental directory.
In this patch we're moving a version of the Tag Processor which
was updated during the Core merge into the `lib/compat` directory
so that we can continue to provide it where necessary while
avoiding conflicts that might otherwise arise by leaving it
where it was.

[55203]: https://core.trac.wordpress.org/changeset/55203