test: Improve mobile editor performance tests (#48101)
* test: Linter considers measurePerfomrance an assertion

Performance tests often only assert the performance of a component via
`measurePerformance`. This change adds `measurePerformance` to the list
of functions that are considered assertions to avoid erroneous lint
warnings.

* test: Fix errors in Rich Text performance test

- Remove unnecessary lint warning disable line caused by erroneous
  placement of `screen` argument.
- Query UI components via user-facing attributes, e.g. a11y label.
- Pass mock function to required `onChange` prop to `RichText`
  component.

* test: Rename native editor performance test files

To avoid web test scripts from including the performance files for the
native mobile editor, we must include the term "native" in the file
name. This renames the test files to follow the existing pattern of not
including "test" in the file name.