Tag Processor: Add bookmark invalidation logic (#47559)
Invalidate bookmarks in a `WP_HTML_Tag_Processor` instance if the tags that they're pointing to have been replaced.

While `WP_HTML_Tag_Processor` currently only supports changing a given tag's attributes, we plan to also provide methods to make broader changes (possibly through a subclass of `WP_HTML_Tag_Processor`).

An API like that will have the potential of replacing a tag that a bookmark points to. As a preparation, we thus need to make sure that all bookmarks affected by a HTML replacement are invalidated (i.e. released).

Co-authored-by: Dennis Snell <dennis.snell@automattic.com>