ResponsiveWrapper: Fixed rendering SVGs as a featured image.
The feature image gets rendered within a ResponsiveWrapper. That relies on the image's height and width being able to be determined. In the PostFeaturedImage component, it isn't able to do that for an SVG. This leads the ResponsiveWrapper assigning an undefined value to the paddingBottom style.

As suggested, I added a check for an undefined value and in that case set padding bottom to 100%.

Props aaronrobertshaw, Mamaduka.
Fixes #48184, #5477.