ColorPanel: Fixes missing contrast failure notice.
Right now the colors to check the contrast are being retrieved by using getComputedStyle. This works for some blocks, but not for all.

We already have access to definedColors, and even do some checks, but it was not being used for contrast checking.

I noticed for the button, where the contrast message was missing, the correct colors were present in this definedColors constant. In this commit I use this constant if it's present, and if not, I use the earlier solution as a fallback.

The notice now shows up for buttons, and also still for paragraphs and the cover block.

Fixes #48308.