[Private APIs] Make the re-registration safeguard opt-in rather than opt-out
Gutenberg introduced a system of sharing private APIs in
WordPress/gutenberg#46131. One of the safeguards is a check preventing
the same module from opting-in twice so that contributors cannot easily
gain access by pretending to be a core module.

That safeguard is only meant for WordPress core and not for
the released @wordpress packages. However, right now it is
opt-out and must be explicitly disabled by developers wanting to
install the @wordpress packages. Let's make it opt-out instead.

This commit makes the check opt-in rather than opt-out. Its counterpart
in the wordpress-develop repo makes WordPress explicitly set
the ALLOW_EXPERIMENT_REREGISTRATION to false:

https://github.com/WordPress/wordpress-develop/pull/4121